+++ src/apps/sequencer/model/Scale.h
@@ -25,6 +25,7 @@ public:
     virtual bool isChromatic() const = 0;
     virtual bool isNotePresent(int note) const = 0;
     virtual int getNoteIndex(int note) const = 0;
+    virtual int noteIndex(int note, int rootNote) const = 0;
 
     virtual void noteName(StringBuilder &str, int note, int rootNote, Format format = Long) const = 0;
     virtual float noteToVolts(int note) const = 0;
@@ -35,6 +36,12 @@ public:
     static int Count;
     static const Scale &get(int index);
     static const char *name(int index);
+
+    bool operator==(const Scale& rhs) const
+    {
+        return (_displayName == rhs._displayName);
+    
+    }
 
 private:
     const char *displayName() const { return _displayName; }
@@ -111,6 +118,21 @@ public:
         }
     }
 
+    int noteIndex(int note, int rootNote) const override {
+        int octave = roundDownDivide(note, _noteCount);
+        int noteIndex = 0;
+        if (isChromatic()) {
+            noteIndex = _notes[note - octave * _noteCount] / 128 + rootNote;
+            while (noteIndex >= 12) {
+                noteIndex -= 12;
+                octave += 1;
+            }
+        } else {
+            noteIndex = note - octave * _noteCount + 1;
+        }
+        return noteIndex;
+    }
+
     float noteToVolts(int note) const override {
         int octave = roundDownDivide(note, _noteCount);
         int index = note - octave * _noteCount;
@@ -165,6 +187,10 @@ public:
     }
 
     int getNoteIndex(int note) const override {
+        return -1;
+    }
+
+    int noteIndex(int note, int rootNote) const override {
         return -1;
     }