+++ src/apps/sequencer/ui/model/NoteSequenceListModel.h
@@ -6,6 +6,7 @@
 
 #include "model/NoteSequence.h"
 #include "model/Scale.h"
+#include <iostream>
 
 class NoteSequenceListModel : public RoutableListModel {
 public:
@@ -21,7 +22,16 @@ public:
     };
 
     NoteSequenceListModel()
-    {}
+    {
+        _scales[0] = -1;
+        for (int i = 1; i < 23; ++i) {
+            _scales[i] = i-1;
+        }
+
+        for (int i = 0; i < 8; ++i) {
+            _selectedScale[i] = 0;
+        }
+    }
 
     void setSequence(NoteSequence *sequence) {
         _sequence = sequence;
@@ -82,6 +92,13 @@ public:
         }
     }
 
+    void setSelectedScale(int defaultScale) {
+        if (_editScale) {
+            _sequence->editScale(_scales[_selectedScale[_sequence->trackIndex()]], false, defaultScale);
+        }
+        _editScale = !_editScale;
+    }
+
 private:
     static const char *itemName(Item item) {
         switch (item) {
@@ -118,8 +135,16 @@ private:
         case ResetMeasure:
             _sequence->printResetMeasure(str);
             break;
-        case Scale:
-            _sequence->printScale(str);
+        case Scale: {
+                int trackIndex = _sequence->trackIndex();
+                bool isRouted = Routing::isRouted(Routing::Target::Scale, trackIndex);
+                if (isRouted) {
+                    _sequence->printScale(str);
+                } else {
+                    auto name = _scales[_selectedScale[trackIndex]] < 0 ? "Default" : Scale::name(_scales[_selectedScale[trackIndex]]);
+                    str(name);
+                }
+            }
             break;
         case RootNote:
             _sequence->printRootNote(str);
@@ -146,8 +171,14 @@ private:
         case ResetMeasure:
             _sequence->editResetMeasure(value, shift);
             break;
-        case Scale:
-            _sequence->editScale(value, shift);
+        case Scale: {
+                int trackIndex = _sequence->trackIndex();
+                bool isRouted = Routing::isRouted(Routing::Target::Scale, trackIndex);
+                if (!isRouted) {
+                    int trackIndex = _sequence->trackIndex();
+                    _selectedScale[trackIndex] = clamp(_selectedScale[trackIndex] + value, 0, 23);
+                }
+            }
             break;
         case RootNote:
             _sequence->editRootNote(value, shift);
@@ -221,4 +252,8 @@ private:
     }
 
     NoteSequence *_sequence;
+    private:
+        std::array<int, 23> _scales;
+        std::array<int, 8> _selectedScale;
+        bool _editScale = false;
 };