+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - random generator: random seed just on init method
 - patter follow
 - pattern chain quick shortcut from pattern page
+- scale edit: scales are changed only if the encoder is pressed
+- on scale change the sequence notes are changed according to the previous scale. if a note in the previous scale is also present in the new scale the value is preeserved. if it is not present the nearest note in the new scale will be selected
 
 ## v0.1.46 (4 January 2024)